+++ src/args.rs
@@ -9,7 +9,7 @@ pub struct Args {
     #[arg(long, help = "Hugging Face tokenizer (defaults to model name)")]
     pub tokenizer: Option<String>,
 
-    #[arg(long, default_value = "2", help = "Number of requests")]
+    #[arg(long, default_value = "10", help = "Number of requests")]
     pub max_num_completed_requests: u32,
 
     #[arg(long, default_value = "1", help = "Parallel requests")]
@@ -18,13 +18,13 @@ pub struct Args {
     #[arg(long, default_value = "550", help = "Target input length")]
     pub mean_input_tokens: u32,
 
-    #[arg(long, default_value = "150", help = "Input length variance")]
+    #[arg(long, default_value = "0", help = "Input length variance")]
     pub stddev_input_tokens: u32,
 
     #[arg(long, help = "Target output length [default: none]")]
     pub mean_output_tokens: Option<u32>,
 
-    #[arg(long, default_value = "10", help = "Output length variance")]
+    #[arg(long, default_value = "0", help = "Output length variance")]
     pub stddev_output_tokens: u32,
 
     #[arg(long, default_value = "result_outputs", help = "Output directory")]